Split test cases for Commits tab. Surpress 'CompressingCodec' for tests those requires segment attrubutes.
