+++ src/test/java/org/apache/lucene/luke/models/commits/CommitsImplTest.java
@@ -17,82 +17,15 @@
 
 package org.apache.lucene.luke.models.commits;
 
-import org.apache.lucene.analysis.MockAnalyzer;
 import org.apache.lucene.codecs.Codec;
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.Field;
-import org.apache.lucene.index.DirectoryReader;
-import org.apache.lucene.index.IndexWriterConfig;
-import org.apache.lucene.index.NoDeletionPolicy;
-import org.apache.lucene.index.RandomIndexWriter;
-import org.apache.lucene.store.Directory;
-import org.apache.lucene.util.LuceneTestCase;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-import java.io.IOException;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 
-public class CommitsImplTest extends LuceneTestCase {
-
-  private DirectoryReader reader;
-
-  private Directory dir;
-
-  private Path indexDir;
-
-  @Override
-  @Before
-  public void setUp() throws Exception {
-    super.setUp();
-    indexDir = createIndex();
-    dir = newFSDirectory(indexDir);
-    reader = DirectoryReader.open(dir);
-  }
-
-  private Path createIndex() throws IOException {
-    Path indexDir = createTempDir();
-
-    Directory dir = newFSDirectory(indexDir);
-
-    IndexWriterConfig config = new IndexWriterConfig(new MockAnalyzer(random()));
-    config.setIndexDeletionPolicy(NoDeletionPolicy.INSTANCE);
-    RandomIndexWriter writer = new RandomIndexWriter(random(), dir, config);
-
-    Document doc1 = new Document();
-    doc1.add(newStringField("f1", "1", Field.Store.NO));
-    writer.addDocument(doc1);
-
-    writer.commit();
-
-    Document doc2 = new Document();
-    doc2.add(newStringField("f1", "2", Field.Store.NO));
-    writer.addDocument(doc2);
-
-    Document doc3 = new Document();
-    doc3.add(newStringField("f1", "3", Field.Store.NO));
-    writer.addDocument(doc3);
-
-    writer.commit();
-
-    writer.close();
-    dir.close();
-
-    return indexDir;
-  }
-
-  @Override
-  @After
-  public void tearDown() throws Exception {
-    super.tearDown();
-    reader.close();
-    dir.close();
-  }
+public class CommitsImplTest extends CommitsTestBase {
 
   @Test
   public void testListCommits() {
@@ -143,28 +76,6 @@ public class CommitsImplTest extends LuceneTestCase {
     CommitsImpl commits = new CommitsImpl(reader, indexDir.toString());
     assertTrue(commits.getSegments(3).isEmpty());
   }
-
-  @Test
-  public void testGetSegmentAttributes() {
-    CommitsImpl commits = new CommitsImpl(reader, indexDir.toString());
-    Map<String, String> attributes = commits.getSegmentAttributes(1, "_0");
-    assertTrue(attributes.size() > 0);
-  }
-
-  @Test
-  public void testGetSegmentAttributes_generation_notfound() {
-    CommitsImpl commits = new CommitsImpl(reader, indexDir.toString());
-    Map<String, String> attributes = commits.getSegmentAttributes(3, "_0");
-    assertTrue(attributes.isEmpty());
-  }
-
-  @Test
-  public void testGetSegmentAttributes_invalid_name() {
-    CommitsImpl commits = new CommitsImpl(reader, indexDir.toString());
-    Map<String, String> attributes = commits.getSegmentAttributes(1, "xxx");
-    assertTrue(attributes.isEmpty());
-  }
-
 
   @Test
   public void testGetSegmentDiagnostics() {
